% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelpuzzle.R
\name{colorize_sprite}
\alias{colorize_sprite}
\title{Colorize a sprite}
\usage{
colorize_sprite(img, colors = c("X", "S", "W", "S", "X", "ORI"), sleep = 0.1)
}
\arguments{
\item{img}{Sprite}

\item{colors}{Vector of colors that are used}

\item{sleep}{Time between setting colors}
}
\value{
Nothing
}
\description{
Colorize a sprite
}
