% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelpuzzle.R
\name{combine_sprite}
\alias{combine_sprite}
\title{Combine two sprite}
\usage{
combine_sprite(img, img2, gap = 1)
}
\arguments{
\item{img}{Sprite 1}

\item{img2}{Sprite 2 (added on the right)}

\item{gap}{Gap between the sprites}
}
\value{
Sprite
}
\description{
Combine two sprite
}
\examples{
sprites <- define_sprite()
sprite1 <- sprites[["buhu"]]
sprite2 <- sprites[["sword"]]
show_sprite(combine_sprite(sprite1, sprite2))
}
