% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelpuzzle.R
\name{define_sprite}
\alias{define_sprite}
\title{Returns a list of predefined sprites}
\usage{
define_sprite()
}
\value{
List of sprites
}
\description{
Returns a list of predefined sprites
}
\examples{
sprites <- define_sprite()

## show first sprite
show_sprite(sprites[[1]])
## show sprite named "sword"
show_sprite(sprites[["sword"]])
}
