% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelpuzzle.R
\name{flip_sprite}
\alias{flip_sprite}
\title{Flip a sprite}
\usage{
flip_sprite(img)
}
\arguments{
\item{img}{Sprite}
}
\value{
Sprite
}
\description{
Flip a sprite
}
\examples{
sprites <- define_sprite()
sprite <- sprites[["sword"]]
show_sprite(flip_sprite(sprite))
}
