% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelpuzzle.R
\name{intro}
\alias{intro}
\title{Intro of pixelpuzzle game}
\usage{
intro(sleep = 0.1)
}
\arguments{
\item{sleep}{How long to wait between change of color}
}
\value{
Nothing
}
\description{
Intro of pixelpuzzle game
}
