% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelpuzzle.R
\name{pixelpuzzle}
\alias{pixelpuzzle}
\title{Play pixelpuzzle in the R-console}
\usage{
pixelpuzzle(img = NA, bg = ".")
}
\arguments{
\item{img}{Image of Sprite (vector of strings). If no sprite is provided,
the player can choose one of the predefined pixel arts.}

\item{bg}{Background color, default is transparent (".")}
}
\value{

}
\description{
Play pixelpuzzle in the R-console
}
\examples{
## Start game (in interactive R sessions)
if (interactive())  {
   pixelpuzzle()
}
}
