% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelpuzzle.R
\name{play_sprite}
\alias{play_sprite}
\title{Play pixelpuzzle with a sprite}
\usage{
play_sprite(img, ori, bg = ".")
}
\arguments{
\item{img}{Sprite to play with}

\item{ori}{Original Sprite}

\item{bg}{Background ("." is transparent)}
}
\value{
Nothing
}
\description{
Play pixelpuzzle with a sprite
}
