% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelpuzzle.R
\name{select_sprite}
\alias{select_sprite}
\title{Select one of the predefined sprites}
\usage{
select_sprite(sprites, bg = ".")
}
\arguments{
\item{sprites}{List of predefined sprites}

\item{bg}{Background of the sprite ("." = transparent)}
}
\value{
Sprite
}
\description{
Select one of the predefined sprites
}
