% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelpuzzle.R
\name{shift_sprite_row}
\alias{shift_sprite_row}
\title{Sift a row of the sprite}
\usage{
shift_sprite_row(img, row = 1, shift = 1)
}
\arguments{
\item{img}{Image of Sprite (vector of strings)}

\item{row}{Which row is shifted}

\item{shift}{Nuber of shifts}
}
\value{
Sprite
}
\description{
Sift a row of the sprite
}
